% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_amortization.R
\name{cf_amortization}
\alias{cf_amortization}
\title{Computing amortization}
\usage{
cf_amortization(r, y, p, M)
}
\arguments{
\item{r}{interest rate}

\item{y}{number of years}

\item{p}{number of periods for year}

\item{M}{loan ammount}
}
\value{
A data.table with amortization information
}
\description{
Function to compute the amortization tables
}
\examples{
r<-0.16
y<-10
p<-12
M<-10000
A<-cf_amortization( r, y, p, M )
sum( A$c )
sum( A$i )
}
\author{
Pedro Guarderas
}
