% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cf_diff_solv_cns}
\alias{cf_diff_solv_cns}
\title{Diffusion solver with Crank-Nicolson scheme}
\usage{
cf_diff_solv_cns(theta, alpha, I, A, B, t, x)
}
\arguments{
\item{theta}{Parameter for the Crank-Nicolson scheme}

\item{alpha}{Diffusion parameter}

\item{I}{Initial condition}

\item{A}{Inferior boundary condition}

\item{B}{Superior boundary condition}

\item{t}{Time grid}

\item{x}{Spatial grid}
}
\value{
List with solution parameters
}
\description{
Solver for diffusion problems implemented with Crank-Nicolson scheme
}
\note{
Diffusion solver for pricing options
}
\author{
Pedro Guarderas
}
