% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cf_diff_solv_euls}
\alias{cf_diff_solv_euls}
\title{Diffusion solver}
\usage{
cf_diff_solv_euls(alpha, I, A, B, t, x)
}
\arguments{
\item{alpha}{Diffusion parameter}

\item{I}{Initial condition}

\item{A}{Inferior boundary condition}

\item{B}{Superior boundary condition}

\item{t}{Time grid}

\item{x}{Spatial grid}
}
\value{
List with solution parameters
}
\description{
Solver for standard diffusion problems
}
\note{
Diffusion solver for pricing options
}
\author{
Pedro Guarderas
}
