% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cf_edo_solv_precor}
\alias{cf_edo_solv_precor}
\title{Ordinary differenatial equation solver}
\usage{
cf_edo_solv_precor(t, v0, f, m = 2L, err = 0.01)
}
\arguments{
\item{t}{time grid, could be adapted}

\item{v0}{initial value}

\item{f}{dynamic function}

\item{m}{number of iterations}

\item{err}{tolerance error}
}
\value{
Return the numerical solution to the ODE.
}
\description{
Solver for ordinary differential equations, implemented with the predictor
corrector method.
}
\note{
The solver is implemented to compute the solution forwards in time.
}
\author{
Pedro Guarderas
\email{pedro.felipe.guarderas@gmail.com}
}
