% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_smoothing_spline.R
\name{cf_forward_rate}
\alias{cf_forward_rate}
\title{Tasa}
\usage{
cf_forward_rate(u, t, b)
}
\arguments{
\item{u}{valor}

\item{t}{tiempo}

\item{b}{coeficientes de interpolación}
}
\value{
vector
}
\description{
Tasa
}
\author{
Pedro Guarderas
}
