% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_lattice.R
\name{cf_lattice}
\alias{cf_lattice}
\title{Multinomial lattice}
\usage{
cf_lattice(N, U, S0)
}
\arguments{
\item{N}{number of periods for the multinomial model}

\item{U}{vector of changes}

\item{S0}{initial price}
}
\value{
A list with a tree structure of the asset evolution
}
\description{
The option pricing can be made with a multinomial lattice, this functions can
generate such a lattice with aid of some give parameters.
}
\author{
Pedro Guarderas
\email{pedro.felipe.guarderas@gmail.com}
}
