% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cf_psor_solv}
\alias{cf_psor_solv}
\title{PSOR algorithm}
\usage{
cf_psor_solv(u0, A, b, c, w, n, e)
}
\arguments{
\item{u0}{initial guest of the solution}

\item{A}{matrix determining the variational inequality}

\item{b}{right term of the inequality}

\item{c}{lower constraint for the solution}

\item{w}{weight}

\item{n}{maximal number of iterations}

\item{e}{relative error for the solution improvement}
}
\description{
Projected successive over-relaxation
}
\author{
Pedro Guarderas
\email{pedro.felipe.guarderas@gmail.com}
}
