% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_smoothing_spline.R
\name{cf_smoothing_spline}
\alias{cf_smoothing_spline}
\title{Spline basado en el método de Lorimier}
\usage{
cf_smoothing_spline(t, y, alpha)
}
\arguments{
\item{t}{tiempo}

\item{y}{rendimiento}

\item{alpha}{factor de ajuste}
}
\value{
lista con resultados de la interpolación, coeficientes y matriz
}
\description{
Spline basado en el método de Lorimier
}
\author{
Pedro Guarderas
}
