% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cf_sor_solv}
\alias{cf_sor_solv}
\title{SOR algorithm}
\usage{
cf_sor_solv(u0, A, b, c, w, n, e)
}
\arguments{
\item{u0}{initial guest of the solution}

\item{A}{matrix determining the variational inequality}

\item{c}{lower constraint for the solution}

\item{w}{weight}

\item{n}{maximal number of iterations}

\item{e}{relative error for the solution improvement}
}
\description{
Solver tridiagonal matrices.
}
\author{
Pedro Guarderas
}
