% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_tree_pricing.R
\name{cf_tree_pricing}
\alias{cf_tree_pricing}
\title{Pricing options with a multinomial tree}
\usage{
cf_tree_pricing(Q, EQ, R, S, option = identity, type = "E", option.par = NULL)
}
\arguments{
\item{Q}{equivalent discrete martingale measure}

\item{EQ}{discrete version of the equivalent discrete martingale average}

\item{R}{term structure of the interest rate, could be a fixed value or a multinomial lattice}

\item{S}{multinomial tree}

\item{option}{function defining the option over S}

\item{type}{option type a character that specifies the king of option, by default 'E' european
option, 'A' american option, 'F' futures option, 'S' swap option, 'P' ...}

\item{option.par}{list of parameter for the option}
}
\value{
A list with a tree structure of the asset evolution
}
\description{
The option pricing can be made with a multinomial tree, this functions can
price different types of options
}
\note{
Pricing a Forward option is like to price an European option with the identity function
regarded as the option
}
\author{
Pedro Guarderas
}
