% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cf_tri_diag_solv}
\alias{cf_tri_diag_solv}
\title{Tridiagonal solver}
\usage{
cf_tri_diag_solv(a, b, c, d)
}
\arguments{
\item{a}{lower diagonal}

\item{b}{diagonal}

\item{c}{upper diagonal}

\item{d}{image of the solution vector, which over written with the solution}
}
\description{
Solver tridiagonal matrices.
}
\author{
Pedro Guarderas
}
