% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cf_uniform_grid}
\alias{cf_uniform_grid}
\title{Uniform grid}
\usage{
cf_uniform_grid(a, b, N)
}
\arguments{
\item{a}{inferior value for the grid}

\item{b}{superior value for the grid}

\item{n}{number of points in the grid}
}
\value{
A vector with the grid points.
}
\description{
simple function for grid generation
}
\author{
Pedro Guarderas
}
