% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cf_wiener}
\alias{cf_wiener}
\title{Brownian motion}
\usage{
cf_wiener(d, t)
}
\arguments{
\item{d}{Dimension}

\item{t}{Time grid}
}
\value{
List with solution parameters
}
\description{
Simulate d-dimensional Browninan motion
}
\note{
Diffusion solver for pricing options
}
\author{
Pedro Guarderas
}
