% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_smoothing_spline.R
\name{cf_yield}
\alias{cf_yield}
\title{Rendimiento}
\usage{
cf_yield(u, b, t)
}
\arguments{
\item{u}{valor}

\item{b}{coeficientes de interpolación}

\item{t}{tiempo}
}
\value{
vector
}
\description{
Rendimiento
}
\author{
Pedro Guarderas
}
