% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_lattice_class.R
\docType{class}
\name{cflattice-class}
\alias{cflattice-class}
\title{Lattice class for multinomial valuation}
\value{
A class of name cflattice
}
\description{
Lattice employed for multinomial valuation
}
\section{Slots}{

\describe{
\item{\code{lattice}}{lattice of values}

\item{\code{order}}{order of the lattice}
}}

\author{
Pedro Guarderas
}
