% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_lattice_class.R
\docType{class}
\name{cflattice-class}
\alias{cflattice-class}
\alias{cflattice,}
\alias{[,cflattice-method}
\alias{[,}
\alias{ANY,}
\alias{ANY-method}
\alias{length,cflattice-method}
\alias{length,}
\alias{numeric,}
\alias{order}
\alias{order,}
\alias{order,cflattice-method}
\title{An S4 class for lattice}
\usage{
\S4method{[}{cflattice}(x, i, j, ..., drop = TRUE)

\S4method{length}{cflattice}(x)

order(x)

\S4method{order}{cflattice}(x)
}
\arguments{
\item{x}{object of class cflattice}

\item{i}{index}

\item{j}{index}

\item{...}{extra parameters}

\item{drop}{drop clause}
}
\value{
Numeric vector with the respective values

Numeric value

Numeric value

Numeric value
}
\description{
Lattice employed for multinomial valuation

Extract method for cflattice

Extract the length of the cflattice

Extract the order of cflattice

Extract the order of cflattice
}
\section{Slots}{

\describe{
\item{\code{lattice}}{lattice of values}

\item{\code{order}}{order of the lattice}
}}

\author{
Pedro Guarderas
\email{pedro.felipe.guarderas@gmail.com}
}
