% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFLatticePrincing.R
\name{CFLatticePricing}
\alias{CFLatticePricing}
\title{Option pricing with the multinomial model}
\usage{
CFLatticePricing(Q, EQ, R, S, option, Type = "E")
}
\arguments{
\item{Q}{equivalent discrete martingale measure}

\item{EQ}{discrete version of the equivalent discrete martingale average}

\item{R}{term structure of the interest rate, could be a fixed value or a multinomial lattice}

\item{S}{multinomial lattice}

\item{option}{function defining the option over S}

\item{Type}{option type a character that specifies the king of option, by default 'E' european
option, 'A' american option, 'F' futures option, 'S' swap option, 'P' ...}
}
\value{
A list with a tree structure of the asset evolution
}
\description{
Function for pricing an option with a multinomial model
}
\examples{
s<-0.3
T<-0.25
N<-15
c<-0.01
u<-exp( s * sqrt( T / N ) )
d<-1/u
r<-0.02
S0<-100
K<-110
R<-exp( r * T / N )
q<-( R - d ) / ( u - d )
Q<-c( 1-q, q )
U<-c( d, u )

# K is global
call<-function( S ) {
  max( S - K, 0 )
}

put<-function( S ) {
  max( K - S, 0 )
}

# Equivalent measure
EQ<-function(R,Q,C) {
  return( sum( R * Q * C ) )
}

S<-CFLattice( N, U, S0 )

# Pricing american call
Ca<-CFLatticePricing( Q, EQ, R, S, call, Type = 'A' )

# Pricing american put
Pa<-CFLatticePricing( Q, EQ, R, S, put, Type = 'A' )
}
\author{
Pedro Guarderas
}
\seealso{
\code{\link{CFLattice}}
}

