% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CFStochSolv}
\alias{CFStochSolv}
\title{Brownian motion}
\usage{
CFStochSolv(d1, d2, X0, b, s, t)
}
\value{
Solution to the stochastic differential equation
}
\description{
Simulate d-dimensional Brownian motion
}
\note{
Diffusion solver for pricing options
}
\author{
Pedro Guarderas
}
