% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CFThieleSolv}
\alias{CFThieleSolv}
\title{Thiele equations solver}
\usage{
CFThieleSolv(t, V0, b, B, theta = 0.5)
}
\arguments{
\item{t}{time grid, could be adapted}

\item{V0}{initial mathematical reserve}

\item{b}{fixed benefits}

\item{B}{transition benefits}
}
\value{
Return a list with the mathematical reserve, the time grid.
}
\description{
Solver for Thiele equation and computation of Mathematical reserves. The solver 
is implemented with a Crank-Nicolson algorithm.
}
\note{
The solver is implemented to compute the solution forwards in time.
}
\author{
Pedro Guarderas
}

