% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFTree.R
\name{CFTree}
\alias{CFTree}
\title{Multinomial tree}
\usage{
CFTree(n, u, S0)
}
\arguments{
\item{n}{number of periods for the multinomial model}

\item{u}{vector of changes}

\item{S0}{initial price}
}
\value{
A list with a tree structure of the asset evolution
}
\description{
The option pricing can be made with a multinomial tree, this functions can 
generate such a lattice with aid of some give parameters.
}
\author{
Pedro Guarderas
}

