% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CFTriDiagSolv}
\alias{CFTriDiagSolv}
\title{Tridiagonal solver}
\usage{
CFTriDiagSolv(a, b, c, d)
}
\arguments{
\item{a}{lower diagonal}

\item{b}{diagonal}

\item{c}{upper diagonal}

\item{d}{image of the solution vector, which over written with the solution}
}
\description{
Solver tridiagonal matrices.
}
\author{
Pedro Guarderas
}
