% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_lattice_class.R
\name{length,cflattice-method}
\alias{length,cflattice-method}
\title{Length method}
\usage{
\S4method{length}{cflattice}(x)
}
\arguments{
\item{x}{object of class cflattice}
}
\value{
Numeric value
}
\description{
Extract the length of the cflattice
}
\author{
Pedro Guarderas
}
