% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultinomialLattice.R
\name{MLattice}
\alias{MLattice}
\title{Multinomial lattice}
\usage{
MLattice(N, U, S0)
}
\arguments{
\item{N}{size of the multinomial model}

\item{U}{vector of probabilities}

\item{S0}{initial price}
}
\value{
A list with a tree structure of the asset evolution
}
\description{
The option pricing can be made with a multinomial lattice, this functions can 
generate such a lattice with aid of some give parameters.
}
\examples{
library( CFINI )
s<-0.3
T<-0.25
N<-15
u<-exp( s * sqrt( T / N ) )
d<-1/u
r<-0.02
S0<-100
R<-exp( r * T / N )
q<-( R - d ) / ( u - d )
U<-c( d, u )
S<-MLattice( N, U, S0 )
}
\author{
Pedro Guarderas
}

