% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultinomialPrincing.R
\name{MultPrice}
\alias{MultPrice}
\title{Option pricing with the multinomial model}
\usage{
MultPrice(S, option, EQ, R, Q, Type = "E")
}
\arguments{
\item{S}{multinomial lattice}

\item{option}{function defining the option over S}

\item{EQ}{discrete version of the equivalent discrete martingale average}

\item{R}{term structure of the interest rate, could be a fixed value or a multinomial lattice}

\item{Q}{equivalent discrete martingale measure}

\item{type}{option type a character that specifies the king of option, by default 'E' european
option, 'A' american option, 'F' futures option, 'S' swap option, 'P' ...}
}
\value{
A list with a tree structure of the asset evolution
}
\description{
Function for pricing an option with a multinomial model
}
\examples{
library( CFINI )
s<-0.3
T<-0.25
N<-15
c<-0.01
u<-exp( s * sqrt( T/N))
d<-1/u
r<-0.02
S0<-100
K<-110
R<-exp( r * T / N )
q<-( R - d ) / ( u - d )
Q<-c( 1-q, q )
U<-c( d, u )

# K is global
call<-function( S ) {
  max( S - K, 0 )
}

put<-function( S ) {
  max( K - S, 0 )
}

# Equivalent measure
EQ<-function(R,Q,C) {
  return( sum( R * Q * C ) )
}

S<-MLattice( N, U, S0 )

# Pricing american call
Ca<-MPrice( S, call, EQ, R, Q, Type = 'A' )

# Pricing american put
Pa<-MPrice( S, put, EQ, R, Q, Type = 'A' )
}
\author{
Pedro Guarderas
}
\seealso{
\code{\link{MLattice}}
}

