% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_lattice_class.R
\name{order,cflattice-method}
\alias{order,cflattice-method}
\title{Order method}
\usage{
\S4method{order}{cflattice}(x)
}
\arguments{
\item{x}{object of class cflattice}
}
\value{
Numeric value
}
\description{
Extract the order of cflattice
}
\author{
Pedro Guarderas
}
