% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_lattice_class.R
\name{[,cflattice,ANY,ANY,ANY-method}
\alias{[,cflattice,ANY,ANY,ANY-method}
\title{Extraction method}
\usage{
\S4method{[}{cflattice,ANY,ANY,ANY}(x, i)
}
\arguments{
\item{x}{object of class cflattice}

\item{i}{index}
}
\value{
Numeric vector with the respective values
}
\description{
Extract method for cflattice
}
\author{
Pedro Guarderas
}
