% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_lattice_class.R
\name{[}
\alias{[}
\title{[ extraction operator}
\usage{
\S4method{[}{cflattice}(x, i)
}
\arguments{
\item{x}{object of class cflattice}

\item{i}{index}
}
\value{
Numeric vector with the respective values
}
\description{
Extract method for cflattice
}
\author{
Pedro Guarderas
\email{pedro.felipe.guarderas@gmail.com}
}
