% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cf_adapt_grid}
\alias{cf_adapt_grid}
\title{Exponential grid}
\usage{
cf_adapt_grid(l, a, b, n, E)
}
\arguments{
\item{l}{adaptation points}

\item{a}{inferior value for the grid}

\item{b}{superior value for the grid}

\item{n}{number of points in the grid}

\item{E}{refinement parameter}
}
\value{
A vector with the adaptive grid points.
}
\description{
Adaptive grid generator
}
\author{
Pedro Guarderas
\email{pedro.felipe.guarderas@gmail.com}
}
