% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cf_black_scholes_solv_cns}
\alias{cf_black_scholes_solv_cns}
\title{Black-Scholes solver}
\usage{
cf_black_scholes_solv_cns(sigma, rate, theta, u0, u1, u2, t, x)
}
\arguments{
\item{sigma}{Volatility}

\item{rate}{Interest rate}

\item{theta}{Parameter for the Crank-Nicolson scheme}

\item{u0}{vector with discretization of the initial condition}

\item{u1}{vector with discretization of the inferior boundary condition}

\item{u2}{vector with discretization of the superior boundary condition}

\item{t}{vector with discretization of the time grid}

\item{x}{vector with discretization of the spatial grid}
}
\value{
List with solution parameters
}
\description{
Solver for Black-Scholes models implemented with the Crank-Nicolson numerical
scheme
}
\note{
pricing options
}
\author{
Pedro Guarderas
}
