% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cf_diff_solv_cns}
\alias{cf_diff_solv_cns}
\title{Diffusion solver with Crank-Nicolson scheme}
\usage{
cf_diff_solv_cns(theta, alpha, u0, u1, u2, t, x, is_initial)
}
\arguments{
\item{theta}{Parameter for the Crank-Nicolson scheme}

\item{alpha}{matrix discretization Diffusion parameter}

\item{u0}{vector with discretization of the initial condition}

\item{u1}{vector with discretization of the inferior boundary condition}

\item{u2}{vector with discretization of the superior boundary condition}

\item{t}{vector with discretization of the time grid}

\item{x}{vector with discretization of the spatial grid}

\item{is_initial}{specifies if the condition is final of initial}
}
\value{
List with solution parameters
}
\description{
Solver for diffusion problems implemented with Crank-Nicolson scheme
}
\note{
Diffusion solver for pricing options
}
\author{
Pedro Guarderas
\email{pedro.felipe.guarderas@gmail.com}
}
