% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cf_diff_solv_euls}
\alias{cf_diff_solv_euls}
\title{Diffusion solver with implicit Euler scheme}
\usage{
cf_diff_solv_euls(alpha, u0, u1, u2, t, x, is_initial)
}
\arguments{
\item{alpha}{matrix discretization Diffusion parameter}

\item{u0}{vector with discretization of the initial condition}

\item{u1}{vector with discretization of the inferior boundary condition}

\item{u2}{vector with discretization of the superior boundary condition}

\item{t}{vector with discretization of the time grid}

\item{x}{vector with discretization of the spatial grid}
}
\value{
List with solution parameters
}
\description{
Solver for diffusion problems implemented using the Euler implicit scheme
}
\note{
Diffusion solver for pricing options
}
\author{
Pedro Guarderas
\email{pedro.felipe.guarderas@gmail.com}
}
