% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CFBlackScholesSolvCNS}
\alias{CFBlackScholesSolvCNS}
\title{Black-Scholes solver}
\usage{
CFBlackScholesSolvCNS(sigma, rate, theta, I, A, B, t, x)
}
\value{
List with solution parameters
}
\description{
Solver for Black-Scholes models implemented with Crank-Nicolson scheme
}
\note{
pricing options
}
\author{
Pedro Guarderas
}
