% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CFDiffSolvCNS}
\alias{CFDiffSolvCNS}
\title{Diffusion solver}
\usage{
CFDiffSolvCNS(alpha, theta, I, A, B, t, x)
}
\value{
List with solution parameters
}
\description{
Solver for diffusion problems implemented with Crank-Nicolson scheme
}
\note{
Diffusion solver for pricing options
}
\author{
Pedro Guarderas
}
