% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CFDiffSolvES}
\alias{CFDiffSolvES}
\title{Diffusion solver}
\usage{
CFDiffSolvES(alpha, I, A, B, t, x)
}
\value{
List with solution parameters
}
\description{
Solver for standard diffusion problems
}
\note{
Diffusion solver for pricing options
}
\author{
Pedro Guarderas
}
