% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFElemenSecurity.R
\name{CFElemSecurity}
\alias{CFElemSecurity}
\title{Elementary securities}
\usage{
CFElemSecurity(n, q, u, r)
}
\arguments{
\item{n}{number of periods for the multinomial model}

\item{q}{neutral risk probabilities}

\item{u}{vector of changes}

\item{r}{interest rate}
}
\value{
A list with a tree structure of the asset evolution
}
\description{
Computes the elementary securities for a multinomial model
}
\author{
Pedro Guarderas
}
