% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_lattice_class.R
\name{order}
\alias{order}
\title{Generic order method}
\usage{
order(x)
}
\arguments{
\item{x}{object of class cflattice}
}
\value{
Numeric value
}
\description{
Extract the order of cflattice
}
\author{
Pedro Guarderas
}
